chore(pin): move pinning to the application controller
